Move SchemaSidebar to schema dir
